+++ root/web/src/main/java/com/synclite/dbreader/web/ValidateDBReader.java
@@ -1756,8 +1756,11 @@ public class ValidateDBReader extends HttpServlet {
 
 				StringBuilder columnBuilder = new StringBuilder();
 				String columnTypeName = columns.getString("TYPE_NAME");
-				int columnSize = columns.getInt("COLUMN_SIZE");
-				int decimalDigits = columns.getInt("DECIMAL_DIGITS");
+				//COLUMN_SIZE / DECIMAL_DIGITS must be read as long: MySQL reports
+				//4294967295 (2^32-1) for types such as INT UNSIGNED, LONGTEXT and
+				//LONGBLOB, which overflows java.lang.Integer and makes getInt throw.
+				long columnSize = columns.getLong("COLUMN_SIZE");
+				long decimalDigits = columns.getLong("DECIMAL_DIGITS");
 				//this.globalTracer.info("column : " + objectName + "." + columnName);
 				String isNullable = getIsNullableClause(columns.getString("IS_NULLABLE")); // NULL or NOT NULL
 				if (columnName.matches(".*\\s+.*")) {
